test(core): refactor O3 failure tests into fuzz tests
