+++ core/src/main/java/io/questdb/cairo/pool/WriterPool.java
@@ -379,6 +379,15 @@ public class WriterPool extends AbstractPool {
             e.owner = UNALLOCATED;
             notifyListener(e.owner, tableToken, PoolListener.EV_CREATE_EX);
             throw ex;
+        } catch (CairoError ex) {
+            LOG.critical().$("could not open [table=`").utf8(tableToken.getTableName())
+                    .$("`, thread=").$(e.owner)
+                    .$(", ex=").utf8(ex.getFlyweightMessage())
+                    .$(']').$();
+            e.ownershipReason = OWNERSHIP_REASON_WRITER_ERROR;
+            e.owner = UNALLOCATED;
+            notifyListener(e.owner, tableToken, PoolListener.EV_CREATE_EX);
+            throw ex;
         }
     }
