+++ core/src/test/java/io/questdb/test/griffin/O3FailureFuzzTest.java
@@ -1,0 +1,513 @@
+/*******************************************************************************
+ *     ___                  _   ____  ____
+ *    / _ \ _   _  ___  ___| |_|  _ \| __ )
+ *   | | | | | | |/ _ \/ __| __| | | |  _ \
+ *   | |_| | |_| |  __/\__ \ |_| |_| | |_) |
+ *    \__\_\\__,_|\___||___/\__|____/|____/
+ *
+ *  Copyright (c) 2014-2019 Appsicle
+ *  Copyright (c) 2019-2023 QuestDB
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ ******************************************************************************/
+
+package io.questdb.test.griffin;
+
+import io.questdb.cairo.CairoEngine;
+import io.questdb.cairo.CairoError;
+import io.questdb.cairo.CairoException;
+import io.questdb.griffin.SqlCompiler;
+import io.questdb.griffin.SqlException;
+import io.questdb.griffin.SqlExecutionContext;
+import io.questdb.std.Chars;
+import io.questdb.std.Files;
+import io.questdb.std.FilesFacade;
+import io.questdb.std.Rnd;
+import io.questdb.std.str.LPSZ;
+import io.questdb.test.std.TestFilesFacadeImpl;
+import io.questdb.test.tools.TestUtils;
+import org.jetbrains.annotations.NotNull;
+import org.junit.After;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
+
+@RunWith(Parameterized.class)
+public class O3FailureFuzzTest extends AbstractO3Test {
+    private final static AtomicInteger counter = new AtomicInteger(0);
+    private final static AtomicBoolean failNextAlloc = new AtomicBoolean(false);
+    private static final TestFilesFacadeImpl ffAllocateFailure = new TestFilesFacadeImpl() {
+
+        @Override
+        public boolean allocate(int fd, long size) {
+            if (failNextAlloc.get()) {
+                failNextAlloc.set(false);
+                return false;
+            }
+            return super.allocate(fd, size);
+        }
+
+        @Override
+        public long length(int fd) {
+            if (counter.decrementAndGet() == 0) {
+                failNextAlloc.set(true);
+                return 0;
+            }
+            return super.length(fd);
+        }
+    };
+    private static final FilesFacade ffOpenFailure = new TestFilesFacadeImpl() {
+        @Override
+        public int openRW(LPSZ name, long opts) {
+            if ((Chars.endsWith(name, Files.SEPARATOR + "ts.d") && Chars.contains(name, "1970-01-06") && counter.decrementAndGet() == 0) && failNextAlloc.get()) {
+                failNextAlloc.set(false);
+                return -1;
+            }
+            return super.openRW(name, opts);
+        }
+    };
+
+    private static int failCounter;
+    private static Rnd rnd;
+    private final int workerCount;
+
+    public O3FailureFuzzTest(ParallelMode mode) {
+        this.workerCount = mode == ParallelMode.Contended ? 0 : 2;
+    }
+
+    @Parameterized.Parameters(name = "{0}")
+    public static Collection<Object[]> data() {
+        return Arrays.asList(new Object[][]{
+                {ParallelMode.Parallel},
+                {ParallelMode.Contended}
+        });
+    }
+
+    public static void reset() {
+        failNextAlloc.set(false);
+        counter.set(Integer.MAX_VALUE);
+    }
+
+    @Before
+    public void setUp() {
+        super.setUp();
+        rnd = TestUtils.generateRandom(LOG);
+    }
+
+    @After
+    public void tearDown() throws Exception {
+        reset();
+        super.tearDown();
+    }
+
+    @Test
+    public void testFuzzAllocationFailure() throws Exception {
+        runFuzzRoutine(O3FailureFuzzTest::testPartitionedDataAppendOODataNotNullStrTailFailRetry0);
+    }
+
+    @Test
+    public void testPartitionedDataAppend() throws Exception {
+        rnd = new Rnd(965901126417583L, 1686331970700L);
+        runFuzzRoutine(O3FailureFuzzTest::testPartitionedDataAppendOOPrependOODatThenRegularAppend0);
+    }
+
+    @Test
+    public void testPartitionedDataAppendOOPrependOOData() throws Exception {
+        runFuzzRoutine(O3FailureFuzzTest::testPartitionedDataAppendOOPrependOODataFailRetry0);
+    }
+
+    @Test
+    public void testPartitionedDataAppendOOPrependOODataMapVarContended() throws Exception {
+        reset();
+        failCounter = rnd.nextInt(5);
+        executeWithPool(0, O3FailureFuzzTest::testPartitionedDataAppendOOPrependOODataFailRetry0, new TestFilesFacadeImpl() {
+
+            @Override
+            public long mmap(int fd, long len, long offset, int flags, int memoryTag) {
+                if (failNextAlloc.get() && this.fd == fd) {
+                    failNextAlloc.set(false);
+                    this.fd = -1;
+                    return -1;
+                }
+                return super.mmap(fd, len, offset, flags, memoryTag);
+            }
+
+            @Override
+            public int openRW(LPSZ name, long opts) {
+                int fd = super.openRW(name, opts);
+                if (Chars.endsWith(name, "1970-01-06" + Files.SEPARATOR + "m.d") && counter.decrementAndGet() == 0) {
+                    this.fd = fd;
+                    failNextAlloc.set(true);
+                }
+                return fd;
+            }
+        });
+    }
+
+    @Test
+    public void testPartitionedOpenTimestampFailContended() throws Exception {
+        runFuzzRoutineOpen(O3FailureFuzzTest::testPartitionedDataAppendOOPrependOODataFailRetry0);
+    }
+
+    private static void assertXCountAndMax(
+            SqlCompiler compiler,
+            SqlExecutionContext sqlExecutionContext,
+            CharSequence expectedMaxTimestamp
+    ) throws SqlException {
+        assertXCount(compiler, sqlExecutionContext);
+        assertMaxTimestamp(compiler.getEngine(), expectedMaxTimestamp);
+    }
+
+    @NotNull
+    private static String prepareCountAndMaxTimestampSinks(SqlCompiler compiler, SqlExecutionContext sqlExecutionContext) throws SqlException {
+        TestUtils.printSql(
+                compiler,
+                sqlExecutionContext,
+                "select count() from x",
+                sink2
+        );
+
+        TestUtils.printSql(
+                compiler,
+                sqlExecutionContext,
+                "select max(ts) from x",
+                sink
+        );
+
+        return Chars.toString(sink);
+    }
+
+    private static void reset(int failCounter) {
+        reset();
+        counter.set(failCounter);
+    }
+
+    private static void testPartitionedDataAppendOODataNotNullStrTailFailRetry0(
+            CairoEngine engine,
+            SqlCompiler compiler,
+            SqlExecutionContext sqlExecutionContext
+    ) throws SqlException {
+        compiler.compile(
+                "create table x as (" +
+                        "select" +
+                        " cast(x as int) i," +
+                        " rnd_symbol('msft','ibm', 'googl') sym," +
+                        " round(rnd_double(0)*100, 3) amt," +
+                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
+                        " rnd_boolean() b," +
+                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
+                        " rnd_double(2) d," +
+                        " rnd_float(2) e," +
+                        " rnd_short(10,1024) f," +
+                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
+                        " rnd_symbol(4,4,4,2) ik," +
+                        " rnd_long() j," +
+                        " timestamp_sequence(500000000000L,100000000L) ts," +
+                        " rnd_byte(2,50) l," +
+                        " cast(null as binary) m," +
+                        " rnd_str(5,16,2) n," +
+                        " rnd_char() t" +
+                        " from long_sequence(510)" +
+                        "), index(sym) timestamp (ts) partition by DAY",
+                sqlExecutionContext
+        );
+
+        compiler.compile(
+                "create table append as (" +
+                        "select" +
+                        " cast(x as int) i," +
+                        " rnd_symbol('msft','ibm', 'googl') sym," +
+                        " round(rnd_double(0)*100, 3) amt," +
+                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
+                        " rnd_boolean() b," +
+                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
+                        " rnd_double(2) d," +
+                        " rnd_float(2) e," +
+                        " rnd_short(10,1024) f," +
+                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
+                        " rnd_symbol(4,4,4,2) ik," +
+                        " rnd_long() j," +
+                        " timestamp_sequence(518300000010L,100000L) ts," +
+                        " rnd_byte(2,50) l," +
+                        " rnd_bin(10, 20, 2) m," +
+                        " rnd_str(5,16,2) n," +
+                        " rnd_char() t" +
+                        " from long_sequence(100)" +
+                        ") timestamp (ts) partition by DAY",
+                sqlExecutionContext
+        );
+
+        final String expectedMaxTimestamp = prepareCountAndMaxTimestampSinks(compiler, sqlExecutionContext);
+
+        for (int i = 0; i < 20; i++) {
+            try {
+                reset(failCounter);
+                compiler.compile("insert into x select * from append", sqlExecutionContext);
+                return;
+            } catch (CairoException | CairoError ignore) {
+            }
+        }
+
+        reset();
+
+        assertXCountAndMax(compiler, sqlExecutionContext, expectedMaxTimestamp);
+
+        assertO3DataConsistency(
+                engine,
+                compiler,
+                sqlExecutionContext,
+                "create table y as (x union all append)",
+                "insert into x select * from append"
+        );
+
+        assertIndexConsistency(compiler, sqlExecutionContext, engine);
+        assertXCountY(compiler, sqlExecutionContext);
+    }
+
+    private static void testPartitionedDataAppendOOPrependOODatThenRegularAppend0(
+            CairoEngine engine,
+            SqlCompiler compiler,
+            SqlExecutionContext sqlExecutionContext
+    ) throws SqlException {
+        // create table with roughly 2AM data
+        compiler.compile(
+                "create table x as (" +
+                        "select" +
+                        " cast(x as int) i," +
+                        " rnd_symbol('msft','ibm', 'googl') sym," +
+                        " round(rnd_double(0)*100, 3) amt," +
+                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
+                        " rnd_boolean() b," +
+                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
+                        " rnd_double(2) d," +
+                        " rnd_float(2) e," +
+                        " rnd_short(10,1024) f," +
+                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
+                        " rnd_symbol(4,4,4,2) ik," +
+                        " rnd_long() j," +
+                        " timestamp_sequence(500000000000L,100000000L) ts," +
+                        " rnd_byte(2,50) l," +
+                        " cast(null as binary) m," +
+                        " rnd_str(5,16,2) n," +
+                        " rnd_char() t" +
+                        " from long_sequence(510)" +
+                        "), index(sym) timestamp (ts) partition by DAY",
+                sqlExecutionContext
+        );
+
+        // all records but one is appended to middle partition
+        // last record is prepended to the last partition
+        compiler.compile(
+                "create table append as (" +
+                        "select" +
+                        " cast(x as int) i," +
+                        " rnd_symbol('msft','ibm', 'googl') sym," +
+                        " round(rnd_double(0)*100, 3) amt," +
+                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
+                        " rnd_boolean() b," +
+                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
+                        " rnd_double(2) d," +
+                        " rnd_float(2) e," +
+                        " rnd_short(10,1024) f," +
+                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
+                        " rnd_symbol(4,4,4,2) ik," +
+                        " rnd_long() j," +
+                        " timestamp_sequence(518390000000L,100000L) ts," +
+                        " rnd_byte(2,50) l," +
+                        " rnd_bin(10, 20, 2) m," +
+                        " rnd_str(5,16,2) n," +
+                        " rnd_char() t" +
+                        " from long_sequence(101)" +
+                        ") timestamp (ts) partition by DAY",
+                sqlExecutionContext
+        );
+
+        final String expectedMaxTimestamp = prepareCountAndMaxTimestampSinks(compiler, sqlExecutionContext);
+
+        for (int i = 0; i < 20; i++) {
+            try {
+                reset(failCounter);
+                compiler.compile("insert into x select * from append", sqlExecutionContext);
+                return;
+            } catch (CairoException ignored) {
+            } catch (CairoError e) {
+                Assert.assertEquals(e.getMessage(), "Table 'x' is distressed");
+            }
+        }
+
+        reset();
+        assertXCountAndMax(compiler, sqlExecutionContext, expectedMaxTimestamp);
+
+        // all records but one is appended to middle partition
+        // last record is prepended to the last partition
+        compiler.compile(
+                "create table append2 as (" +
+                        "select" +
+                        " cast(x as int) i," +
+                        " rnd_symbol('msft','ibm', 'googl') sym," +
+                        " round(rnd_double(0)*100, 3) amt," +
+                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
+                        " rnd_boolean() b," +
+                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
+                        " rnd_double(2) d," +
+                        " rnd_float(2) e," +
+                        " rnd_short(10,1024) f," +
+                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
+                        " rnd_symbol(4,4,4,2) ik," +
+                        " rnd_long() j," +
+                        " timestamp_sequence(551000000000L,100000L) ts," +
+                        " rnd_byte(2,50) l," +
+                        " rnd_bin(10, 20, 2) m," +
+                        " rnd_str(5,16,2) n," +
+                        " rnd_char() t" +
+                        " from long_sequence(101)" +
+                        ") timestamp (ts) partition by DAY",
+                sqlExecutionContext
+        );
+
+
+        // create third table, which will contain both X and 1AM
+        assertO3DataConsistency(
+                engine,
+                compiler,
+                sqlExecutionContext,
+                "create table y as (x union all append2)",
+                "insert into x select * from append2"
+        );
+
+        assertIndexConsistency(
+                compiler,
+                sqlExecutionContext,
+                engine);
+
+        assertXCountY(compiler, sqlExecutionContext);
+    }
+
+    private static void testPartitionedDataAppendOOPrependOODataFailRetry0(
+            CairoEngine engine,
+            SqlCompiler compiler,
+            SqlExecutionContext sqlExecutionContext
+    ) throws SqlException {
+        // create table with roughly 2AM data
+        compiler.compile(
+                "create table x as (" +
+                        "select" +
+                        " cast(x as int) i," +
+                        " rnd_symbol('msft','ibm', 'googl') sym," +
+                        " round(rnd_double(0)*100, 3) amt," +
+                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
+                        " rnd_boolean() b," +
+                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
+                        " rnd_double(2) d," +
+                        " rnd_float(2) e," +
+                        " rnd_short(10,1024) f," +
+                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
+                        " rnd_symbol(4,4,4,2) ik," +
+                        " rnd_long() j," +
+                        " timestamp_sequence(500000000000L,100000000L) ts," +
+                        " rnd_byte(2,50) l," +
+                        " cast(null as binary) m," +
+                        " rnd_str(5,16,2) n," +
+                        " rnd_char() t" +
+                        " from long_sequence(510)" +
+                        "), index(sym) timestamp (ts) partition by DAY",
+                sqlExecutionContext
+        );
+
+        // all records but one is appended to middle partition
+        // last record is prepended to the last partition
+        compiler.compile(
+                "create table append as (" +
+                        "select" +
+                        " cast(x as int) i," +
+                        " rnd_symbol('msft','ibm', 'googl') sym," +
+                        " round(rnd_double(0)*100, 3) amt," +
+                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
+                        " rnd_boolean() b," +
+                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
+                        " rnd_double(2) d," +
+                        " rnd_float(2) e," +
+                        " rnd_short(10,1024) f," +
+                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
+                        " rnd_symbol(4,4,4,2) ik," +
+                        " rnd_long() j," +
+                        " timestamp_sequence(518390000000L,100000L) ts," +
+                        " rnd_byte(2,50) l," +
+                        " rnd_bin(10, 20, 2) m," +
+                        " rnd_str(5,16,2) n," +
+                        " rnd_char() t" +
+                        " from long_sequence(101)" +
+                        ") timestamp (ts) partition by DAY",
+                sqlExecutionContext
+        );
+
+        final String expectedMaxTimestamp = prepareCountAndMaxTimestampSinks(compiler, sqlExecutionContext);
+
+        for (int i = 0; i < 20; i++) {
+            try {
+                reset(failCounter);
+                compiler.compile("insert into x select * from append", sqlExecutionContext);
+                return;
+            } catch (CairoException ignored) {
+            } catch (CairoError e) {
+                Assert.assertEquals(e.getMessage(), "Table 'x' is distressed");
+            }
+        }
+
+        reset();
+
+        assertXCountAndMax(compiler, sqlExecutionContext, expectedMaxTimestamp);
+
+        if (rnd.nextBoolean()) {
+            engine.releaseAllWriters();
+        }
+
+        // create third table, which will contain both X and 1AM
+        assertO3DataConsistency(
+                engine,
+                compiler,
+                sqlExecutionContext,
+                "create table y as (x union all append)",
+                "insert into x select * from append"
+        );
+
+        assertIndexConsistency(
+                compiler,
+                sqlExecutionContext,
+                engine);
+
+        assertXCountY(compiler, sqlExecutionContext);
+    }
+
+    private void runFuzzRoutine(CustomisableRunnable routine) throws Exception {
+        reset();
+        failCounter = 10 + rnd.nextInt(200);
+        LOG.info().$("failCounter=").$(failCounter).$();
+        executeWithPool(workerCount, routine, ffAllocateFailure);
+    }
+
+    private void runFuzzRoutineOpen(CustomisableRunnable routine) throws Exception {
+        reset();
+        failCounter = rnd.nextInt(26);
+        LOG.info().$("failCounter=").$(failCounter).$();
+        executeWithPool(workerCount, routine, ffOpenFailure);
+    }
+}
