+++ core/src/test/java/io/questdb/test/griffin/O3FailureTest.java
@@ -70,120 +70,6 @@ public class O3FailureTest extends AbstractO3Test {
         }
     };
     private final static AtomicBoolean fixFailure = new AtomicBoolean(true);
-    private static final FilesFacade ffAllocateFailure = new TestFilesFacadeImpl() {
-        private final AtomicInteger increment = new AtomicInteger();
-        private boolean failNextAlloc = false;
-
-        @Override
-        public boolean allocate(int fd, long size) {
-            if (!fixFailure.get() || failNextAlloc) {
-                failNextAlloc = false;
-                fixFailure.set(false);
-                return false;
-            }
-            return super.allocate(fd, size);
-        }
-
-        @Override
-        public long length(int fd) {
-            if (fd > 0) {
-                final long remaining = counter.decrementAndGet();
-                if (!fixFailure.get() || remaining == 0) {
-                    failNextAlloc = true;
-                    return 0;
-                }
-            } else {
-                // For debugging, if new call with ff.length(fd) is added, change it
-                // to ff.length(-fd) and adjust counter set by the test by the number
-                // in increment when 0 is returned
-                increment.incrementAndGet();
-            }
-            return super.length(Math.abs(fd));
-        }
-    };
-    private static final FilesFacade ffFailToAllocateIndex = new TestFilesFacadeImpl() {
-        boolean failNextAlloc = false;
-        int theFd;
-
-        @Override
-        public boolean allocate(int fd, long size) {
-            if (!fixFailure.get() || (fd == theFd && failNextAlloc)) {
-                failNextAlloc = false;
-                return false;
-            }
-            return super.allocate(fd, size);
-        }
-
-        @Override
-        public boolean close(int fd) {
-            if (fd == theFd) {
-                theFd = -1;
-            }
-            return super.close(fd);
-        }
-
-        @Override
-        public long length(int fd) {
-            if (!fixFailure.get() || (fd == theFd && counter.decrementAndGet() == 0)) {
-                fixFailure.set(false);
-                failNextAlloc = true;
-                return 0L;
-            }
-            return super.length(fd);
-        }
-
-        @Override
-        public int openRW(LPSZ name, long opts) {
-            int fd = super.openRW(name, opts);
-            if (Chars.endsWith(name, "x" + TableUtils.SYSTEM_TABLE_NAME_SUFFIX + Files.SEPARATOR + "1970-01-07" + Files.SEPARATOR + "m.i")) {
-                theFd = fd;
-            }
-            return fd;
-        }
-    };
-    private static final FilesFacade ffIndexAllocateFailure = new TestFilesFacadeImpl() {
-        boolean failNextAlloc = false;
-        int theFd = 0;
-
-        @Override
-        public boolean allocate(int fd, long size) {
-            if (!fixFailure.get() || (fd == theFd && failNextAlloc)) {
-                // don't forget to set this to 0 so that next attempt doesn't fail
-                theFd = 0;
-                failNextAlloc = false;
-                fixFailure.set(false);
-                return false;
-            }
-            return super.allocate(fd, size);
-        }
-
-        @Override
-        public boolean close(int fd) {
-            if (fd > 0 && fd == theFd) {
-                theFd = 0;
-            }
-            return super.close(fd);
-        }
-
-        @Override
-        public long length(int fd) {
-            if (!fixFailure.get() || fd == theFd) {
-                failNextAlloc = true;
-                fixFailure.set(false);
-                return 0;
-            }
-            return super.length(fd);
-        }
-
-        @Override
-        public int openRW(LPSZ name, long opts) {
-            int fd = super.openRW(name, opts);
-            if (Chars.endsWith(name, "1970-01-06" + Files.SEPARATOR + "sym.v") && counter.decrementAndGet() == 0) {
-                theFd = fd;
-            }
-            return fd;
-        }
-    };
     private static final FilesFacade ffMapRW = new TestFilesFacadeImpl() {
 
         @Override
@@ -221,16 +107,6 @@ public class O3FailureTest extends AbstractO3Test {
                 return -1;
             }
             return super.mkdirs(path, mode);
-        }
-    };
-    private static final FilesFacade ffOpenFailure = new TestFilesFacadeImpl() {
-        @Override
-        public int openRW(LPSZ name, long opts) {
-            if (!fixFailure.get() || (Chars.endsWith(name, Files.SEPARATOR + "ts.d") && Chars.contains(name, "1970-01-06") && counter.decrementAndGet() == 0)) {
-                fixFailure.set(false);
-                return -1;
-            }
-            return super.openRW(name, opts);
         }
     };
     private static final FilesFacade ffOpenRW = new TestFilesFacadeImpl() {
@@ -680,13 +556,6 @@ public class O3FailureTest extends AbstractO3Test {
     }
 
     @Test
-    public void testFailOnResizingIndexContended() throws Exception {
-        // this places break point on resize of key file
-        counter.set(152 + 12 + 20 + 19 + 8);
-        executeWithPool(0, O3FailureTest::testPartitionedDataAppendOODataNotNullStrTailFailRetry0, ffAllocateFailure);
-    }
-
-    @Test
     public void testFailOnTruncateKeyIndexContended() throws Exception {
         counter.set(0);
         executeWithPool(0, O3FailureTest::testColumnTopLastOOOPrefixFailRetry0, new TestFilesFacadeImpl() {
@@ -883,18 +752,6 @@ public class O3FailureTest extends AbstractO3Test {
     }
 
     @Test
-    public void testPartitionedAllocateLastPartitionFail() throws Exception {
-        counter.set(2);
-        executeWithoutPool(O3FailureTest::testPartitionedDataAppendOOPrependOODataFailRetry0, ffFailToAllocateIndex);
-    }
-
-    @Test
-    public void testPartitionedAllocateLastPartitionFailNoReopen() throws Exception {
-        counter.set(2);
-        executeWithPool(0, O3FailureTest::testPartitionedDataAppendOOPrependOODataFailRetryNoReopen, ffFailToAllocateIndex);
-    }
-
-    @Test
     public void testPartitionedCreateDirFail() throws Exception {
         counter.set(1);
         executeWithoutPool(O3FailureTest::testColumnTopMidMergeBlankColumnFailRetry0, ffMkDirFailure);
@@ -1007,126 +864,6 @@ public class O3FailureTest extends AbstractO3Test {
     }
 
     @Test
-    public void testPartitionedDataAppendOODataNotNullStrTail() throws Exception {
-        counter.set(174 + 12 + 19 + 8);
-        executeWithoutPool(O3FailureTest::testPartitionedDataAppendOODataNotNullStrTailFailRetry0, ffAllocateFailure);
-    }
-
-    @Test
-    public void testPartitionedDataAppendOODataNotNullStrTailContended() throws Exception {
-        counter.set(174 + 12 + 19 + 8);
-        executeWithPool(0, O3FailureTest::testPartitionedDataAppendOODataNotNullStrTailFailRetry0, ffAllocateFailure);
-    }
-
-    @Test
-    public void testPartitionedDataAppendOODataNotNullStrTailIndexAllocateFail() throws Exception {
-        counter.set(2);
-        executeWithoutPool(O3FailureTest::testPartitionedDataAppendOODataNotNullStrTailFailRetry0, ffIndexAllocateFailure);
-    }
-
-    @Test
-    public void testPartitionedDataAppendOODataNotNullStrTailIndexAllocateFailContended() throws Exception {
-        counter.set(2);
-        executeWithPool(0, O3FailureTest::testPartitionedDataAppendOODataNotNullStrTailFailRetry0, ffIndexAllocateFailure);
-    }
-
-    @Test
-    public void testPartitionedDataAppendOODataNotNullStrTailParallel() throws Exception {
-        counter.set(174 + 45 + 12 + 27 + 12);
-        executeWithPool(2, O3FailureTest::testPartitionedDataAppendOODataNotNullStrTailFailRetry0, ffAllocateFailure);
-    }
-
-    @Test
-    public void testPartitionedDataAppendOOPrependOODatThenRegularAppend() throws Exception {
-        counter.set(165 + 45 + 12 + 21 + 9);
-        executeWithPool(0, O3FailureTest::testPartitionedDataAppendOOPrependOODatThenRegularAppend0, ffAllocateFailure);
-    }
-
-    @Test
-    public void testPartitionedDataAppendOOPrependOOData() throws Exception {
-        counter.set(165 + 45 + 12 + 18 + 8);
-        executeWithoutPool(O3FailureTest::testPartitionedDataAppendOOPrependOODataFailRetry0, ffAllocateFailure);
-    }
-
-    @Test
-    public void testPartitionedDataAppendOOPrependOODataContended() throws Exception {
-        counter.set(165 + 45 + 12 + 21 + 9);
-        executeWithPool(0, O3FailureTest::testPartitionedDataAppendOOPrependOODataFailRetry0, ffAllocateFailure);
-    }
-
-    @Test
-    public void testPartitionedDataAppendOOPrependOODataMapVar() throws Exception {
-        counter.set(3);
-        executeWithoutPool(O3FailureTest::testPartitionedDataAppendOOPrependOODataFailRetry0, new TestFilesFacadeImpl() {
-
-            @Override
-            public boolean close(int fd) {
-                if (fd > 0 && fd == this.fd) {
-                    this.fd = -1;
-                }
-                return super.close(fd);
-            }
-
-            @Override
-            public long mmap(int fd, long len, long offset, int flags, int memoryTag) {
-                if (!fixFailure.get() || this.fd == fd) {
-                    fixFailure.set(false);
-                    this.fd = -1;
-                    return -1;
-                }
-                return super.mmap(fd, len, offset, flags, memoryTag);
-            }
-
-            @Override
-            public int openRW(LPSZ name, long opts) {
-                int fd = super.openRW(name, opts);
-                if (Chars.endsWith(name, "1970-01-06" + Files.SEPARATOR + "m.d") && counter.decrementAndGet() == 0) {
-                    this.fd = fd;
-                }
-                return fd;
-            }
-        });
-    }
-
-    @Test
-    public void testPartitionedDataAppendOOPrependOODataMapVarContended() throws Exception {
-        counter.set(3);
-        executeWithPool(0, O3FailureTest::testPartitionedDataAppendOOPrependOODataFailRetry0, new TestFilesFacadeImpl() {
-
-            @Override
-            public long mmap(int fd, long len, long offset, int flags, int memoryTag) {
-                if (!fixFailure.get() || this.fd == fd) {
-                    fixFailure.set(false);
-                    this.fd = -1;
-                    return -1;
-                }
-                return super.mmap(fd, len, offset, flags, memoryTag);
-            }
-
-            @Override
-            public int openRW(LPSZ name, long opts) {
-                int fd = super.openRW(name, opts);
-                if (Chars.endsWith(name, "1970-01-06" + Files.SEPARATOR + "m.d") && counter.decrementAndGet() == 0) {
-                    this.fd = fd;
-                }
-                return fd;
-            }
-        });
-    }
-
-    @Test
-    public void testPartitionedDataAppendOOPrependOODataParallel() throws Exception {
-        counter.set(193 + 45 + 18 + 27 + 12);
-        executeWithPool(4, O3FailureTest::testPartitionedDataAppendOOPrependOODataFailRetry0, ffAllocateFailure);
-    }
-
-    @Test
-    public void testPartitionedDataAppendOOPrependOODataParallelNoReopen() throws Exception {
-        counter.set(176 + 45 + 18 + 26 + 12);
-        executeWithPool(4, O3FailureTest::testPartitionedDataAppendOOPrependOODataFailRetryNoReopen, ffAllocateFailure);
-    }
-
-    @Test
     public void testPartitionedOOPrefixesExistingPartitions() throws Exception {
         counter.set(1);
         executeWithoutPool(O3FailureTest::testPartitionedOOPrefixesExistingPartitionsFailRetry0, ffOpenIndexFailure);
@@ -1167,18 +904,6 @@ public class O3FailureTest extends AbstractO3Test {
     }
 
     @Test
-    public void testPartitionedOpenTimestampFail() throws Exception {
-        counter.set(3);
-        executeWithoutPool(O3FailureTest::testPartitionedDataAppendOOPrependOODataFailRetry0, ffOpenFailure);
-    }
-
-    @Test
-    public void testPartitionedOpenTimestampFailContended() throws Exception {
-        counter.set(3);
-        executeWithPool(0, O3FailureTest::testPartitionedDataAppendOOPrependOODataFailRetry0, ffOpenFailure);
-    }
-
-    @Test
     public void testPartitionedWithAllocationCallLimit() throws Exception {
         counter.set(0);
         executeWithPool(0, O3FailureTest::testPartitionedWithAllocationCallLimit0, new TestFilesFacadeImpl() {
@@ -1193,12 +918,6 @@ public class O3FailureTest extends AbstractO3Test {
                 return super.allocate(fd, size);
             }
         });
-    }
-
-    @Test
-    public void testSetAppendPositionFails() throws Exception {
-        counter.set(169 + 12 + 20 + 19 + 8);
-        executeWithoutPool(O3FailureTest::testPartitionedDataAppendOODataNotNullStrTailFailRetry0, ffAllocateFailure);
     }
 
     @Test
@@ -2554,8 +2273,7 @@ public class O3FailureTest extends AbstractO3Test {
                         " rnd_str() v11," +
                         " rnd_bin() v12," +
                         " rnd_long() v9" +
-                        " from long_sequence(1000)" +
-                        "",
+                        " from long_sequence(1000)",
                 sqlExecutionContext
         );
 
@@ -3593,324 +3311,6 @@ public class O3FailureTest extends AbstractO3Test {
         );
 
         assertXCountY(compiler, sqlExecutionContext);
-    }
-
-    private static void testPartitionedDataAppendOODataNotNullStrTailFailRetry0(
-            CairoEngine engine,
-            SqlCompiler compiler,
-            SqlExecutionContext sqlExecutionContext
-    ) throws SqlException {
-        compiler.compile(
-                "create table x as (" +
-                        "select" +
-                        " cast(x as int) i," +
-                        " rnd_symbol('msft','ibm', 'googl') sym," +
-                        " round(rnd_double(0)*100, 3) amt," +
-                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
-                        " rnd_boolean() b," +
-                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
-                        " rnd_double(2) d," +
-                        " rnd_float(2) e," +
-                        " rnd_short(10,1024) f," +
-                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
-                        " rnd_symbol(4,4,4,2) ik," +
-                        " rnd_long() j," +
-                        " timestamp_sequence(500000000000L,100000000L) ts," +
-                        " rnd_byte(2,50) l," +
-                        " cast(null as binary) m," +
-                        " rnd_str(5,16,2) n," +
-                        " rnd_char() t" +
-                        " from long_sequence(510)" +
-                        "), index(sym) timestamp (ts) partition by DAY",
-                sqlExecutionContext
-        );
-
-        compiler.compile(
-                "create table append as (" +
-                        "select" +
-                        " cast(x as int) i," +
-                        " rnd_symbol('msft','ibm', 'googl') sym," +
-                        " round(rnd_double(0)*100, 3) amt," +
-                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
-                        " rnd_boolean() b," +
-                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
-                        " rnd_double(2) d," +
-                        " rnd_float(2) e," +
-                        " rnd_short(10,1024) f," +
-                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
-                        " rnd_symbol(4,4,4,2) ik," +
-                        " rnd_long() j," +
-                        " timestamp_sequence(518300000010L,100000L) ts," +
-                        " rnd_byte(2,50) l," +
-                        " rnd_bin(10, 20, 2) m," +
-                        " rnd_str(5,16,2) n," +
-                        " rnd_char() t" +
-                        " from long_sequence(100)" +
-                        ") timestamp (ts) partition by DAY",
-                sqlExecutionContext
-        );
-
-        final String expectedMaxTimestamp = prepareCountAndMaxTimestampSinks(compiler, sqlExecutionContext);
-
-        for (int i = 0; i < 20; i++) {
-            try {
-                compiler.compile("insert into x select * from append", sqlExecutionContext);
-                Assert.fail();
-            } catch (CairoException | CairoError ignore) {
-            }
-        }
-
-        fixFailure.set(true);
-
-        assertXCountAndMax(compiler, sqlExecutionContext, expectedMaxTimestamp);
-
-        assertO3DataConsistency(
-                engine,
-                compiler,
-                sqlExecutionContext,
-                "create table y as (x union all append)",
-                "insert into x select * from append"
-        );
-
-        assertIndexConsistency(compiler, sqlExecutionContext, engine);
-        assertXCountY(compiler, sqlExecutionContext);
-    }
-
-    private static void testPartitionedDataAppendOOPrependOODatThenRegularAppend0(
-            CairoEngine engine,
-            SqlCompiler compiler,
-            SqlExecutionContext sqlExecutionContext
-    ) throws SqlException {
-        // create table with roughly 2AM data
-        compiler.compile(
-                "create table x as (" +
-                        "select" +
-                        " cast(x as int) i," +
-                        " rnd_symbol('msft','ibm', 'googl') sym," +
-                        " round(rnd_double(0)*100, 3) amt," +
-                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
-                        " rnd_boolean() b," +
-                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
-                        " rnd_double(2) d," +
-                        " rnd_float(2) e," +
-                        " rnd_short(10,1024) f," +
-                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
-                        " rnd_symbol(4,4,4,2) ik," +
-                        " rnd_long() j," +
-                        " timestamp_sequence(500000000000L,100000000L) ts," +
-                        " rnd_byte(2,50) l," +
-                        " cast(null as binary) m," +
-                        " rnd_str(5,16,2) n," +
-                        " rnd_char() t" +
-                        " from long_sequence(510)" +
-                        "), index(sym) timestamp (ts) partition by DAY",
-                sqlExecutionContext
-        );
-
-        // all records but one is appended to middle partition
-        // last record is prepended to the last partition
-        compiler.compile(
-                "create table append as (" +
-                        "select" +
-                        " cast(x as int) i," +
-                        " rnd_symbol('msft','ibm', 'googl') sym," +
-                        " round(rnd_double(0)*100, 3) amt," +
-                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
-                        " rnd_boolean() b," +
-                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
-                        " rnd_double(2) d," +
-                        " rnd_float(2) e," +
-                        " rnd_short(10,1024) f," +
-                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
-                        " rnd_symbol(4,4,4,2) ik," +
-                        " rnd_long() j," +
-                        " timestamp_sequence(518390000000L,100000L) ts," +
-                        " rnd_byte(2,50) l," +
-                        " rnd_bin(10, 20, 2) m," +
-                        " rnd_str(5,16,2) n," +
-                        " rnd_char() t" +
-                        " from long_sequence(101)" +
-                        ") timestamp (ts) partition by DAY",
-                sqlExecutionContext
-        );
-
-        final String expectedMaxTimestamp = prepareCountAndMaxTimestampSinks(compiler, sqlExecutionContext);
-
-        for (int i = 0; i < 10; i++) {
-            try {
-                compiler.compile("insert into x select * from append", sqlExecutionContext);
-                Assert.fail();
-            } catch (CairoException ignored) {
-            }
-        }
-
-        fixFailure.set(true);
-
-        assertXCountAndMax(compiler, sqlExecutionContext, expectedMaxTimestamp);
-
-        // all records but one is appended to middle partition
-        // last record is prepended to the last partition
-        compiler.compile(
-                "create table append2 as (" +
-                        "select" +
-                        " cast(x as int) i," +
-                        " rnd_symbol('msft','ibm', 'googl') sym," +
-                        " round(rnd_double(0)*100, 3) amt," +
-                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
-                        " rnd_boolean() b," +
-                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
-                        " rnd_double(2) d," +
-                        " rnd_float(2) e," +
-                        " rnd_short(10,1024) f," +
-                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
-                        " rnd_symbol(4,4,4,2) ik," +
-                        " rnd_long() j," +
-                        " timestamp_sequence(551000000000L,100000L) ts," +
-                        " rnd_byte(2,50) l," +
-                        " rnd_bin(10, 20, 2) m," +
-                        " rnd_str(5,16,2) n," +
-                        " rnd_char() t" +
-                        " from long_sequence(101)" +
-                        ") timestamp (ts) partition by DAY",
-                sqlExecutionContext
-        );
-
-
-        // create third table, which will contain both X and 1AM
-        assertO3DataConsistency(
-                engine,
-                compiler,
-                sqlExecutionContext,
-                "create table y as (x union all append2)",
-                "insert into x select * from append2"
-        );
-
-        assertIndexConsistency(
-                compiler,
-                sqlExecutionContext,
-                engine);
-
-        assertXCountY(compiler, sqlExecutionContext);
-    }
-
-    private static void testPartitionedDataAppendOOPrependOODataFailRetry0(
-            CairoEngine engine,
-            SqlCompiler compiler,
-            SqlExecutionContext sqlExecutionContext
-    ) throws SqlException {
-        testPartitionedDataAppendOOPrependOODataFailRetry0(
-                engine,
-                compiler,
-                sqlExecutionContext,
-                true
-        );
-    }
-
-    private static void testPartitionedDataAppendOOPrependOODataFailRetry0(
-            CairoEngine engine,
-            SqlCompiler compiler,
-            SqlExecutionContext sqlExecutionContext,
-            boolean reopenTableWriter
-    ) throws SqlException {
-        // create table with roughly 2AM data
-        compiler.compile(
-                "create table x as (" +
-                        "select" +
-                        " cast(x as int) i," +
-                        " rnd_symbol('msft','ibm', 'googl') sym," +
-                        " round(rnd_double(0)*100, 3) amt," +
-                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
-                        " rnd_boolean() b," +
-                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
-                        " rnd_double(2) d," +
-                        " rnd_float(2) e," +
-                        " rnd_short(10,1024) f," +
-                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
-                        " rnd_symbol(4,4,4,2) ik," +
-                        " rnd_long() j," +
-                        " timestamp_sequence(500000000000L,100000000L) ts," +
-                        " rnd_byte(2,50) l," +
-                        " cast(null as binary) m," +
-                        " rnd_str(5,16,2) n," +
-                        " rnd_char() t" +
-                        " from long_sequence(510)" +
-                        "), index(sym) timestamp (ts) partition by DAY",
-                sqlExecutionContext
-        );
-
-        // all records but one is appended to middle partition
-        // last record is prepended to the last partition
-        compiler.compile(
-                "create table append as (" +
-                        "select" +
-                        " cast(x as int) i," +
-                        " rnd_symbol('msft','ibm', 'googl') sym," +
-                        " round(rnd_double(0)*100, 3) amt," +
-                        " to_timestamp('2018-01', 'yyyy-MM') + x * 720000000 timestamp," +
-                        " rnd_boolean() b," +
-                        " rnd_str('ABC', 'CDE', null, 'XYZ') c," +
-                        " rnd_double(2) d," +
-                        " rnd_float(2) e," +
-                        " rnd_short(10,1024) f," +
-                        " rnd_date(to_date('2015', 'yyyy'), to_date('2016', 'yyyy'), 2) g," +
-                        " rnd_symbol(4,4,4,2) ik," +
-                        " rnd_long() j," +
-                        " timestamp_sequence(518390000000L,100000L) ts," +
-                        " rnd_byte(2,50) l," +
-                        " rnd_bin(10, 20, 2) m," +
-                        " rnd_str(5,16,2) n," +
-                        " rnd_char() t" +
-                        " from long_sequence(101)" +
-                        ") timestamp (ts) partition by DAY",
-                sqlExecutionContext
-        );
-
-        final String expectedMaxTimestamp = prepareCountAndMaxTimestampSinks(compiler, sqlExecutionContext);
-
-        for (int i = 0; i < 15; i++) {
-            try {
-                compiler.compile("insert into x select * from append", sqlExecutionContext);
-                Assert.fail();
-            } catch (CairoException ignored) {
-            }
-        }
-
-        fixFailure.set(true);
-
-        assertXCountAndMax(compiler, sqlExecutionContext, expectedMaxTimestamp);
-
-        if (reopenTableWriter) {
-            engine.releaseAllWriters();
-        }
-
-        // create third table, which will contain both X and 1AM
-        assertO3DataConsistency(
-                engine,
-                compiler,
-                sqlExecutionContext,
-                "create table y as (x union all append)",
-                "insert into x select * from append"
-        );
-
-        assertIndexConsistency(
-                compiler,
-                sqlExecutionContext,
-                engine);
-
-        assertXCountY(compiler, sqlExecutionContext);
-    }
-
-    private static void testPartitionedDataAppendOOPrependOODataFailRetryNoReopen(
-            CairoEngine engine,
-            SqlCompiler compiler,
-            SqlExecutionContext sqlExecutionContext
-    ) throws SqlException {
-        testPartitionedDataAppendOOPrependOODataFailRetry0(
-                engine,
-                compiler,
-                sqlExecutionContext,
-                false
-        );
     }
 
     private static void testPartitionedOOPrefixesExistingPartitionsFailRetry0(
